// client/user/src/components/button/button_types.ts
import { SerializedStyles } from '@emotion/react';

export enum BUTTON_THEME {
  PRIAMRY = 'PRIAMRY',
  LINEBLUE = 'LINEBLUE',
  LINEBLACK = 'LINEBLACK',
}

export enum BUTTON_SIZE {
  SMALL = 'SMALL',
  MEDIUM = 'MEDIUM',
  LARGE = 'LARGE',
}

export interface IButton {
  label: string;
  type: 'button' | 'submit';
  onClick: () => void;
  isDisabled: boolean;
  theme: BUTTON_THEME;
  size: BUTTON_SIZE;
  styleBlock: boolean;
}

export interface IButtonStyled {
  buttonThemeStyles: SerializedStyles;
  buttonSizeStyles: SerializedStyles;
  styleBlock: boolean;
}

export interface ITextButton {
  label: string;
  onClick: () => void;
  isDisabled: boolean;
  leftIcon?: string;
  rightiCon?: string;
  size: BUTTON_SIZE;
}

export interface ITextButtonStyled {
  buttonSizeStyles: SerializedStyles;
}